+++ src/main/java/finalmission/meetingroom/common/JwtTokenHandler.java
@@ -1,4 +1,4 @@
-package finalmission.meetingroom.service;
+package finalmission.meetingroom.common;
 
 import java.util.Arrays;
 import java.util.Date;
