+++ src/main/java/finalmission/meetingroom/config/WebMvcConfig.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import finalmission.meetingroom.controller.LoginMemberArgumentResolver;
+import finalmission.meetingroom.common.LoginMemberArgumentResolver;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
